Moved padding utilities
